requirements.txt
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
et_xmlfile==2.0.0
h11==0.16.0
idna==3.10
numpy==2.2.5
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
XlsxWriter==3.2.3
